refactor: centralize guardian decision actor context construction
